start outlining Pinpad interface
